moved cast functions to system_wasm
